+++ src/session.rs
@@ -4,27 +4,27 @@ use std::collections::VecDeque;
 
 use eindir_core::{DifferentiableObjective, Objective};
 use ndarray::{Array1, Array2};
-use rand::rngs::StdRng;
 use rand::SeedableRng;
+use rand::rngs::StdRng;
 
-use crate::accept::{accept_step, Accept};
+use crate::accept::{Accept, accept_step};
 use crate::adam::adam_direction;
 use crate::bb::bb_direction;
 use crate::control::Control;
 use crate::error::{Error, Result};
-use crate::fire::{fire_after_v1, fire_displacement, FireState};
+use crate::fire::{FireState, fire_after_v1, fire_displacement};
 use crate::lbfgs::{GradNorm, Lbfgs};
 use crate::linesearch::LineSearch;
 use crate::manifold::{Manifold, ManifoldKind};
 use crate::method::Method;
-use crate::newton::{rfo_direction, shifted_newton, HessianObjective, NewtonKind};
+use crate::newton::{HessianObjective, NewtonKind, rfo_direction, shifted_newton};
 use crate::nlcg::{Conjugacy, ConjugacyContext, Restart};
-use crate::pso::{random_velocity, update_swarm, Particle, RNG_SEED};
+use crate::pso::{Particle, RNG_SEED, random_velocity, update_swarm};
 use crate::qn::{bfgs_inverse_update, solve_dense, sr1_inverse_update, sr2_hessian_update};
 use crate::qn_step::QnStep;
 use crate::report::Report;
 use crate::rigid::{project_horizontal, project_out_rot_trans};
-use crate::step::{l2, next_istep, scale_step, scale_step_atom, take_step};
+use crate::step::{l2, next_istep, qn_istep, scale_step, scale_step_atom, take_step};
 use crate::trust::{
     accept_ratio, dogleg_direction, predicted_reduction, reduction_ratio, update_radius,
 };
@@ -639,7 +639,7 @@ impl Solver {
                 let ev = obj.value_and_gradient(x.view());
                 value = ev.0;
                 grad = ev.1;
-                self.istep = next_istep(lsstep, &self.control);
+                self.istep = qn_istep(&self.control);
             }
             Inner::Nlcg {
                 conjugacy,
@@ -680,7 +680,7 @@ impl Solver {
                 *dir = Array1::from_iter(grad.iter().zip(d_old.iter()).map(|(g, d)| -g + beta * d));
                 g_old.assign(&grad);
                 d_old.assign(dir);
-                self.istep = next_istep(lsstep, &self.control);
+                self.istep = qn_istep(&self.control);
             }
             Inner::Bfgs { h } => {
                 let direction = -h.dot(&grad);
@@ -702,7 +702,7 @@ impl Solver {
                 if moved {
                     bfgs_inverse_update(h, &(&*x - &old), &(&grad - &gold));
                 }
-                self.istep = next_istep(lsstep, &self.control);
+                self.istep = qn_istep(&self.control);
             }
             Inner::Sr1 { h } => {
                 let direction = -h.dot(&grad);